+++ src/db/repositories/list-repository.ts
@@ -64,3 +64,14 @@ export async function insertList(input: InsertListModel) {
     .values({ ...input })
     .returning()
 }
+
+export async function deleteList(id: string) {
+  return db.delete(schema.lists).where(eq(schema.lists.id, id))
+}
+
+export async function getList(id: string, userId: string) {
+  return db
+    .select()
+    .from(schema.lists)
+    .where(and(eq(schema.lists.id, id), eq(schema.lists.userId, userId)))
+}